+++ src/contexts/AlertContext.ts
@@ -1,4 +1,5 @@
 import { createContext } from 'react';
+
 import { IAlertContext } from '../types/alert';
 
 export const AlertContext = createContext<IAlertContext>({
