+++ src/hooks/useAlert.ts
@@ -1,5 +1,7 @@
 import { useContext } from 'react';
+
 import { IAlertContext } from '../types/alert';
+
 import { AlertContext } from '../contexts/AlertContext';
 
 export const useAlert = (): IAlertContext => {
