+++ src/styles/Globalstyle.tsx
@@ -1,0 +1,11 @@
+import { createGlobalStyle } from 'styled-components';
+
+export const GlobalStyle = createGlobalStyle`
+  *{
+    box-sizing:border-box;
+    font-family: 'Pretendard';
+  }
+  body{
+    margin:0;
+  }
+`;
